added syscall class and started finding out how to extract syscall infos out of ctf
