+++ dataloader/recording_ctf.py
@@ -1,4 +1,3 @@
-# from util.bt2.trace_collection_message_iterator import TraceCollectionMessageIterator
 import bt2
 
 from typing import Generator
@@ -16,14 +15,19 @@ class RecordingCTF(BaseRecording):
     def syscalls(self) -> Generator[Syscall, None, None]:
         for msg in self.message_iterator:
             if type(msg) is bt2._EventMessageConst:
-                print(msg.event.name)
+                name = msg.event.name
+                thread_id = msg.event['tid']
+                process_id = msg.event['pid']
+                process_name = msg.event['procname']
+                # timestamp = msg.event['timestamp']
+                yield msg.event.__dict__
 
     def _collect_metadata(self):
         pass
 
 
 if __name__ == '__main__':
-    recording = RecordingCTF("/home/felix/repos/uni/work/LID-DS/scenarios/CVE-2017-7529/CVE-2017-7529_LTTng_CTF_sample/normal_and_attack/defeated_chandrasekhar_3584")
+    recording = RecordingCTF("/home/felix/datasets/CVE-2017-7529_LTTng_CTF_sample/test/normal_and_attack/defeated_chandrasekhar_3584")
 
     for syscall in recording.syscalls():
         print(syscall)
